player/diagnostics: Moved to session/diagnostics.
