+++ server/player/handler.go
@@ -5,8 +5,8 @@ import (
 	"github.com/df-mc/dragonfly/server/cmd"
 	"github.com/df-mc/dragonfly/server/event"
 	"github.com/df-mc/dragonfly/server/item"
-	"github.com/df-mc/dragonfly/server/player/diagnostics"
 	"github.com/df-mc/dragonfly/server/player/skin"
+	"github.com/df-mc/dragonfly/server/session"
 	"github.com/df-mc/dragonfly/server/world"
 	"github.com/go-gl/mathgl/mgl64"
 	"net"
@@ -142,7 +142,7 @@ type Handler interface {
 	// HandleDiagnostics handles the latest diagnostics data that the player has sent to the server. This is
 	// not sent by every client however, only those with the "Creator > Enable Client Diagnostics" setting
 	// enabled.
-	HandleDiagnostics(p *Player, d diagnostics.Diagnostics)
+	HandleDiagnostics(p *Player, d session.Diagnostics)
 }
 
 // NopHandler implements the Handler interface but does not execute any code when an event is called. The
@@ -186,4 +186,4 @@ func (NopHandler) HandleFoodLoss(*Context, int, *int)                           
 func (NopHandler) HandleDeath(*Player, world.DamageSource, *bool)                       {}
 func (NopHandler) HandleRespawn(*Player, *mgl64.Vec3, **world.World)                    {}
 func (NopHandler) HandleQuit(*Player)                                                   {}
-func (NopHandler) HandleDiagnostics(*Player, diagnostics.Diagnostics)                   {}
+func (NopHandler) HandleDiagnostics(*Player, session.Diagnostics)                       {}
