+++ server/session/controllable.go
@@ -7,7 +7,6 @@ import (
 	"github.com/df-mc/dragonfly/server/item"
 	"github.com/df-mc/dragonfly/server/item/inventory"
 	"github.com/df-mc/dragonfly/server/player/chat"
-	"github.com/df-mc/dragonfly/server/player/diagnostics"
 	"github.com/df-mc/dragonfly/server/player/form"
 	"github.com/df-mc/dragonfly/server/player/skin"
 	"github.com/df-mc/dragonfly/server/world"
@@ -107,5 +106,5 @@ type Controllable interface {
 	Skin() skin.Skin
 	SetSkin(skin.Skin)
 
-	UpdateDiagnostics(diagnostics.Diagnostics)
+	UpdateDiagnostics(Diagnostics)
 }
